Multi-capability and svc2svc support for services
This changeset includes updates to enable each IntersectService
instance to provide service for a set of capabilities, rather
than just one. From the client perspective, the key difference
is that operation names should be prepended with the advertised
name of the capability providing the operation.

Also included is support for service-to-service requests, as
is necessary to enable making external requests to other
services/capabilities as part of the implementation of any
service methods. An additional thread is used to process
these external requests asynchronously from incoming request
handlers.